Change Long to Double

+++ api/src/main/java/org/apache/cloudstack/api/response/HostResponse.java
@@ -116,7 +116,7 @@ public class HostResponse extends BaseResponse {
 
     @SerializedName("averageload")
     @Param(description = "the cpu average load on the host")
-    private Long averageLoad;
+    private Double averageLoad;
 
     @SerializedName("networkkbsread")
     @Param(description = "the incoming network traffic on the host")
@@ -337,7 +337,7 @@ public class HostResponse extends BaseResponse {
         this.cpuUsed = cpuUsed;
     }
 
-    public void setCpuAverageLoad(Long averageLoad) {
+    public void setCpuAverageLoad(Double averageLoad) {
         this.averageLoad = averageLoad;
     }
